Add user management guide to Payara Cloud docs
A new guide on how to manage users in Payara Cloud has been added. This guide covers different roles (Administrator, Billing Manager, Member) and scenarios of team management. This is to help users understand their access and privileges within the platform more clearly.

The navigation document has been updated to include a link to the new guide on managing users.